Added support for custom instance methods
